i_img_*_new: avoid signed integer overflow
The test for this was failing for i_img_8_new on MacOS and FreeBSD,
reporting out of memory instead of the overflow.

Any recent perl is built with -fwrapv, but freebsd and macos seem to
drop that, I suspect the check here is getting invalidated with signed
math, try unsigned math.

+++ img16.c
@@ -184,8 +184,8 @@ im_img_16_new(pIMCTX, i_img_dim x, i_img_dim y, int ch) {
     im_push_errorf(aIMCTX, 0, "channels must be between 1 and %d", MAXCHANNELS);
     return NULL;
   }
-  bytes =  x * y * ch * 2;
-  if (bytes / y / ch / 2 != (size_t)x) {
+  bytes =  (size_t)x * (size_t)y * (size_t)ch * 2u;
+  if (bytes / (size_t)y / (size_t)ch / 2u != (size_t)x) {
     im_push_errorf(aIMCTX, 0, "integer overflow calculating image allocation");
     return NULL;
   }
